requirements.txt
alembic==1.9.2
aniso8601==9.0.1
async-timeout==4.0.2
attrs==22.2.0
BareNecessities==0.2.8
bcrypt==3.2.2
beautifulsoup4==4.12.2
blinker==1.6.2
cachelib==0.9.0
cachetools==5.3.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
decorator==5.1.1
Deprecated==1.2.13
distlib==0.3.6
exceptiongroup==1.1.0
filelock==3.9.0
Flask==2.2.2
Flask-Admin==1.6.1
Flask-Bcrypt==1.0.1
Flask-Caching==2.0.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-Limiter==3.3.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==4.0.2
flask-paginate==2022.1.8
flask-restx==1.0.5
Flask-Share==0.1.1
Flask-SQLAlchemy==3.0.2
Flask-UUID==0.2
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.18.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
greenlet==2.0.1
gunicorn==20.1.0
httplib2==0.22.0
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.12.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
ldap3==2.9.1
limits==3.5.0
Mail==2.1.0
Mako==1.2.4
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-sqlalchemy==0.29.0
mdurl==0.1.2
oauthlib==3.2.2
ordered-set==4.1.0
packaging==23.0
platformdirs==2.6.2
pluggy==1.0.0
protobuf==4.23.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.15.1
PyJWT==2.6.0
pyparsing==3.0.9
pypng==0.20220715.0
pyrsistent==0.19.3
pytest==7.2.1
python-decouple==3.7
python-dotenv==0.21.0
pytz==2022.7.1
qrcode==7.4.2
redis==4.5.5
requests==2.30.0
requests-oauthlib==1.3.1
rich==13.3.5
rsa==4.9
six==1.16.0
soupsieve==2.4.1
SQLAlchemy==2.0.15
tomli==2.0.1
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
validators==0.20.0
virtualenv==20.17.1
Werkzeug==2.2.2
wrapt==1.15.0
WTForms==3.0.1
zipp==3.11.0
